Towers of Hanoi for 3 pegs
/*
The Tower of Hanoi is a mathematical game or puzzle. 
It consists of THREE rods and a number of disks of different sizes, which can slide onto any rod. 
The puzzle starts with the disks in a neat stack in ascending order of size on one rod, the smallest at the top, thus making a conical shape.

The objective of the puzzle is to move the entire stack to another rod, obeying the following simple rules:

1. Only one disk can be moved at a time.
2. Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on an empty rod.
3. No disk may be placed on top of a smaller disk.

Minimum number of moves is 2^n - 1 where n is the number of disks.

// Example:
This is for a 3 disk problem.
move disc 1 A to c ACB
move disc 2 A to b ABC
move disc 1 C to b CBA
move disc 3 A to c ACB
move disc 1 B to a BAC
move disc 2 B to c BCA
move disc 1 A to c ACB
Done in 7 moves.
*/